- Fixed a number of bugs/malfunctions in how residents are added to and   moved between residences. - Addresses (residences) can now only be registered for guardians of   juveniles. - Origo members lists now exclude juveniles who are not peers of user   wards. - Added support for adding additional guardians for juveniles, post-   registration. - Added support for toggling gender. - Lists of members to pick from are not passed explicitly to   OValuePickerViewController. - Fixed bug whereby canceling out of member action sheet would trigger edit   mode. - Fixed bug whereby member edits were not saved because a local lookup was   done that reverted all edits. - Fixed bug whereby phone calls could not be placed due to phone number   containing unrecognised formatting characters. - Fixed bug whereby an attempt was made to place multiple phone calls   simultaneously. - Fixed bug whereby list cell would remain selected upon popping containing   view controller. - Fixed bug whereby zombie entities would be persisted when canceling out   of registration views. - Fixed but whereby member would be considered unknown by user if peer of   ward of user but not peer of user himself. - Fixed but whereby guardians would be sorted arbitrarily if one of them   had no address. - Fixed [OOrigoProxy members] so that it lists intended members during   origo registration. - [OOrigo(OrigoAdditions) residence] now always returns residence with   address, if one exists. - Started implementing support for properly handling origo contacts. - Started to reduce dependency on OLanguage language constructs, which   ultimately are bound to fall short of requirements. - Additonal string changes.

+++ OrigoApp/UITableView+OrigoAdditions.m
@@ -62,7 +62,7 @@ static NSString * const kLogoText = @"..origo..";
 
 - (id)listCellWithStyle:(UITableViewCellStyle)style data:(id)data delegate:(id)delegate
 {
-    OTableViewCell *cell = [self cellWithStyle:style reuseIdentifier:[kReuseIdentifierList stringByAppendingFormat:@":%ld", style] delegate:delegate];
+    OTableViewCell *cell = [self cellWithStyle:style reuseIdentifier:[kReuseIdentifierList stringByAppendingFormat:@":%d", (short)style] delegate:delegate];
     
     if ([data conformsToProtocol:@protocol(OEntity)]) {
         cell.entity = data;
